pyproject.toml
[tool.poetry]
name        = "concordia"
version     = "0.1.0"
description = "CONCORDIA – genome-annotation concordance engine"
authors     = ["Your Name <you@anl.gov>"]
license     = "MIT"
readme      = "README.md"

# Tell Poetry to package the code in the `concord/` directory
packages = [{ include = "concord" }]

[tool.poetry.scripts]
concord = "concord.cli:app"        # CLI entry-point  →  `concord …`

[tool.poetry.dependencies]
python                 = "^3.10"

# Core libs
pandas                 = "^2.2"
sentence-transformers  = "^2.7"     # pulls torch, transformers
httpx                  = "^0.27"
typer                  = "^0.9"
pyyaml                 = "^6.0"
tqdm                   = "^4.65.0"
numpy                  = "^1.25.0"
scikit-learn           = "^1.2"
matplotlib             = "^3.7"
seaborn                = "^0.13.1"
torch                  = { version = "^2.2", optional = true }
rich = "^14.0.0"
python-dotenv = "^1.1.0"

[tool.poetry.extras]
gpu = ["torch"]

[tool.poetry.dev-dependencies]
pytest = "^7.4"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.black]
line-length = 88
target-version = ['py310']